Move TimeoutStackStyle to avoid including expensive Channel.hpp every time Settings are needed.
