+++ src/common/Channel.hpp
@@ -20,14 +20,6 @@ namespace chatterino {
 struct Message;
 using MessagePtr = std::shared_ptr<const Message>;
 
-enum class TimeoutStackStyle : int {
-    StackHard = 0,
-    DontStackBeyondUserMessage = 1,
-    DontStack = 2,
-
-    Default = DontStackBeyondUserMessage,
-};
-
 class Channel : public std::enable_shared_from_this<Channel>, public MessageSink
 {
 public:
